Beta

+++ userge/plugins/misc/upload.py
@@ -224,6 +224,8 @@ async def doc_upload(message: Message, path, del_path: bool = False, extra: str 
             progress=progress,
             progress_args=(message, f"uploading {extra}", str(path.name))
         )
+    except ValueError as e_e:
+        await sent.edit(f"Skipping `{path}` due to {e_e}")
     except Exception as u_e:
         await sent.edit(u_e)
         raise u_e
@@ -258,6 +260,8 @@ async def vid_upload(message: Message, path, del_path: bool = False, extra: str 
             progress=progress,
             progress_args=(message, f"uploading {extra}", str(path.name))
         )
+    except ValueError as e_e:
+        await sent.edit(f"Skipping `{path}` due to {e_e}")
     except Exception as u_e:
         await sent.edit(u_e)
         raise u_e
@@ -314,6 +318,8 @@ async def audio_upload(message: Message, path, del_path: bool = False, extra: st
             progress=progress,
             progress_args=(message, f"uploading {extra}", str(path.name))
         )
+    except ValueError as e_e:
+        await sent.edit(f"Skipping `{path}` due to {e_e}")
     except Exception as u_e:
         await sent.edit(u_e)
         raise u_e
@@ -343,6 +349,8 @@ async def photo_upload(message: Message, path, del_path: bool = False, extra: st
             progress=progress,
             progress_args=(message, f"uploading {extra}", str(path.name))
         )
+    except ValueError as e_e:
+        await sent.edit(f"Skipping `{path}` due to {e_e}")
     except Exception as u_e:
         await sent.edit(u_e)
         raise u_e
@@ -358,13 +366,16 @@ async def get_thumb(path: str = ''):
     if os.path.exists(Config.THUMB_PATH):
         return Config.THUMB_PATH
     if path:
+        types = (".jpg", ".webp", ".png")
+        if path.endswith(types):
+            return None
         file_name = os.path.splitext(path)[0]
-        for type_ in (".jpg", ".webp", ".png"):
+        for type_ in types:
             thumb_path = file_name + type_
             if os.path.exists(thumb_path):
                 if type_ != ".jpg":
                     new_thumb_path = f"{file_name}.jpg"
-                    Image.open(thumb_path).save(new_thumb_path, "JPEG")
+                    Image.open(thumb_path).convert('RGB').save(new_thumb_path, "JPEG")
                     os.remove(thumb_path)
                     thumb_path = new_thumb_path
                 return thumb_path
